sender/requirements.txt
dicttoxml==1.7.4
pycryptodome==3.7.2
python-decouple==3.1
requests==2.21.0
